Refactored code from frobots_rigs to fubars, as Fubars.Frobot.Cpu, also the frobots_rigs application/directory changed to frobots, the intent being that it will hold the lua code loading/saving and blockchain interaction logic.
BIG change: added CPU cycles to frobots VMs so that optionally we can interrupt the frobot if they execute too many instructions in one cycle.
Present behaviour is that they experience a cpu overload and lose the current state and is reset to previous state (though the tank may have changed location).